Add ability to define cores
Relying on a known and fixed core count and cores.ratio does not
work well in a heterogenous cluster environment. Adding cores
parameter as a directly setable variable allows for a user to submit
a job through a cluster schedule (Slurm) and set the cores count to
the requested number of cores on the compute node.